master
commit. Unfinished so sorry...

+++ src/week4/com/patikadev/view/OperatorGUI.java
@@ -2,6 +2,8 @@ package week4.com.patikadev.view;
 
 import week4.com.patikadev.helper.Config;
 import week4.com.patikadev.helper.Helper;
+import week4.com.patikadev.helper.Item;
+import week4.com.patikadev.model.Course;
 import week4.com.patikadev.model.Operator;
 import week4.com.patikadev.model.Patika;
 import week4.com.patikadev.model.User;
@@ -10,14 +12,13 @@ import javax.swing.*;
 import javax.swing.event.TableModelEvent;
 import javax.swing.table.DefaultTableModel;
 import java.awt.*;
-import java.awt.event.MouseAdapter;
-import java.awt.event.MouseEvent;
+import java.awt.event.*;
 import java.util.ArrayList;
 import java.util.Objects;
 
 public class OperatorGUI extends JFrame {
 
-    private final Operator operator;
+    private final User operator;
     private JPanel wrapper;
     private JTabbedPane tab_operator;
     private JLabel lbl_welcome;
@@ -42,13 +43,26 @@ public class OperatorGUI extends JFrame {
     private JTextField fld_patika_name;
     private JPanel pnl_patika_add;
     private JButton btn_add_patika;
+    private JPanel pnl_user_list;
+    private JPanel pnl_course_list;
+    private JScrollPane scrl_course_list;
+    private JTable tbl_course_list;
+    private JPanel pnl__course_add;
+    private JTextField fld_course_name;
+    private JTextField fld_course_prog_lang;
+    private JComboBox cmb_course_patika;
+    private JComboBox cmb_course_user;
+    private JButton btn_course_add;
     private DefaultTableModel mdl_user_list;
     private Object[] row_user_list;
     private DefaultTableModel mdl_patika_list;
     private Object[] row_patika_list;
     private JPopupMenu patikaMenu;
+    private DefaultTableModel mdl_course_list;
+    private Object[] row_course_list;
+    private JPopupMenu courseMenu;
 
-    public OperatorGUI(final Operator operator) {
+    public OperatorGUI(final User operator) {
         this.operator = operator;
         add(wrapper);
         setSize(1000, 500);
@@ -72,7 +86,6 @@ public class OperatorGUI extends JFrame {
         Object[] col_user_list = {"ID", "Ad Soyad", "Kullanıcı Adı", "Şifre", "Üyelik Tipi"};
         mdl_user_list.setColumnIdentifiers(col_user_list);
         row_user_list = new Object[col_user_list.length];
-        loadUserModel();
 
         tbl_user_list.getModel().addTableModelListener(e -> {
             if (e.getType() == TableModelEvent.UPDATE) {
@@ -110,8 +123,10 @@ public class OperatorGUI extends JFrame {
             if (Helper.isFieldEmpty(fld_deleteuser_userid)) {
                 Helper.showErrorMessage("Kullanıcı ID'si alanı boş bırakılamaz");
             } else {
-                User.deleteUserByUserId(Integer.valueOf(fld_deleteuser_userid.getText()));
-                loadUserModel();
+                if (Helper.showConfirmMessage("Bu işlem geri alınamaz, emin misiniz? ")) {
+                    User.deleteUserByUserId(Integer.valueOf(fld_deleteuser_userid.getText()));
+                    loadUserModel();
+                }
             }
         });
 
@@ -150,10 +165,38 @@ public class OperatorGUI extends JFrame {
         patikaMenu.add(updateMenu);
         patikaMenu.add(deleteMenu);
 
+        updateMenu.addActionListener(e -> {
+            int selected_row_id = Integer.parseInt(tbl_patika_list.getValueAt(tbl_patika_list.getSelectedRow(), 0).toString());
+            Patika patika = Patika.fetchPatika(selected_row_id);
+            if (patika != null) {
+                UpdatePatikaGUI updatePatikaGUI = new UpdatePatikaGUI(patika);
+                updatePatikaGUI.addWindowListener(new WindowAdapter() {
+                    @Override
+                    public void windowClosed(WindowEvent e) {
+                        loadPatikaModel();
+                    }
+                });
+            } else {
+                Helper.showErrorMessage("Bir sorun meydana geldi. ");
+            }
+        });
+
+        deleteMenu.addActionListener(e -> {
+            int selected_row_id = Integer.parseInt(tbl_patika_list.getValueAt(tbl_patika_list.getSelectedRow(), 0).toString());
+            Patika patika = Patika.fetchPatika(selected_row_id);
+            if (patika != null) {
+                if (Helper.showConfirmMessage("Bu işlem geri alınamaz, kabul ediyor musunuz? ")) {
+                    Patika.delete(selected_row_id);
+                }
+            } else {
+                Helper.showErrorMessage("Bir sorun meydana geldi. ");
+            }
+            loadPatikaModel();
+        });
+
         Object[] col_patika_list = {"ID", "Patika Adı"};
         mdl_patika_list.setColumnIdentifiers(col_patika_list);
         row_patika_list = new Object[col_patika_list.length];
-        loadPatikaModel();
 
         tbl_patika_list.setModel(mdl_patika_list);
         tbl_patika_list.getTableHeader().setReorderingAllowed(false);
@@ -179,6 +222,33 @@ public class OperatorGUI extends JFrame {
             }
             loadPatikaModel();
         });
+
+        // CourseList
+
+        mdl_course_list = new DefaultTableModel();
+        Object[] col_course_list = {"ID", "Ders Adı", "Programlama Dili", "Patika", "Eğitmen"};
+        mdl_course_list.setColumnIdentifiers(col_course_list);
+        row_course_list = new Object[col_course_list.length];
+
+        tbl_course_list.setModel(mdl_course_list);
+        tbl_course_list.getColumnModel().getColumn(0).setMaxWidth(75);
+        tbl_course_list.getTableHeader().setReorderingAllowed(false);
+
+        loadCourseList();
+        loadPatikaModel();
+        loadUserModel();
+        btn_course_add.addActionListener(e -> {
+            Item patikateItem = (Item) cmb_course_patika.getSelectedItem();
+            Item userItem = (Item) cmb_course_user.getSelectedItem();
+            if (Helper.isFieldEmpty(fld_course_name) || Helper.isFieldEmpty(fld_course_prog_lang)) {
+                Helper.showErrorMessage("Lütfen alanları doldurunuz");
+            } else {
+                String courseName = fld_course_name.getText();
+                String progLag = fld_course_prog_lang.getText();
+                Course.add(courseName, progLag, patikateItem.getKey(), userItem.getKey());
+                loadCourseList();
+            }
+        });
     }
 
     public static void main(String[] args) {
@@ -192,13 +262,40 @@ public class OperatorGUI extends JFrame {
         OperatorGUI opGUI = new OperatorGUI(op);
     }
 
+    private void loadPatikaCombo() {
+        cmb_course_patika.removeAllItems();
+        for (Patika patika : Patika.getList()) {
+            cmb_course_patika.addItem(new Item(patika.getId(), patika.getName()));
+        }
+    }
+
+    private void loadUserCombo() {
+        cmb_course_user.removeAllItems();
+        for (User user : User.getList()) {
+            if (user.getType().equals("educator")) {
+                cmb_course_user.addItem(new Item(user.getId(), user.getName()));
+            }
+        }
+    }
+
+    private void loadCourseList() {
+        DefaultTableModel clearTable = (DefaultTableModel) tbl_course_list.getModel();
+        clearTable.setRowCount(0);
+
+        for (Course course : Course.getList()) {
+            toRowItem(course);
+        }
+    }
+
     private void loadPatikaModel() {
         DefaultTableModel clearTable = (DefaultTableModel) tbl_patika_list.getModel();
         clearTable.setRowCount(0);
 
         for (Patika patika : Patika.getList()) {
-            patikaToRowItem(patika);
+            toRowItem(patika);
         }
+
+        loadPatikaCombo();
     }
 
     private void loadUserModel() {
@@ -206,11 +303,13 @@ public class OperatorGUI extends JFrame {
         clearTable.setRowCount(0);
 
         for (User user : User.getList()) {
-            userToRowItem(user);
+            toRowItem(user);
         }
 
         tbl_user_list.setModel(mdl_user_list);
         tbl_user_list.getTableHeader().setReorderingAllowed(false);
+
+        loadUserCombo();
     }
 
     private void loadUserModel(ArrayList<User> userList) {
@@ -218,21 +317,32 @@ public class OperatorGUI extends JFrame {
         clearTable.setRowCount(0);
 
         for (User user : userList) {
-            userToRowItem(user);
+            toRowItem(user);
         }
 
         tbl_user_list.setModel(mdl_user_list);
         tbl_user_list.getTableHeader().setReorderingAllowed(false);
     }
 
-    private void patikaToRowItem(Patika patika) {
+
+    private void toRowItem(Course course) {
+        int i = 0;
+        row_course_list[i++] = course.getId();
+        row_course_list[i++] = course.getName();
+        row_course_list[i++] = course.getLang();
+        row_course_list[i++] = course.getPatika().getName();
+        row_course_list[i++] = course.getEducator().getName();
+        mdl_course_list.addRow(row_course_list);
+    }
+
+    private void toRowItem(Patika patika) {
         int i = 0;
         row_patika_list[i++] = patika.getId();
         row_patika_list[i++] = patika.getName();
         mdl_patika_list.addRow(row_patika_list);
     }
 
-    private void userToRowItem(User user) {
+    private void toRowItem(User user) {
         int i = 0;
         row_user_list[i++] = user.getId();
         row_user_list[i++] = user.getName();
